file upload in nodejs

+++ controllers/admin.js
@@ -1,6 +1,6 @@
 // const {  mongoose } = require('mongoose');
 const Product = require('../models/product');
-
+const fileHelper =require('../util/file')
 exports.getAddProduct = (req, res, next) => {
   
   res.render('admin/edit-product', {
@@ -12,17 +12,19 @@ exports.getAddProduct = (req, res, next) => {
 };
 
 exports.postAddProduct = (req, res, next) => {
-  
   const title = req.body.title;
-  const imageUrl = req.body.imageUrl;
+  const imageUrl = req.file;
   const price = req.body.price;
   const description = req.body.description;
+  if(!imageUrl){
+    throw  new Error("Invalid FILE")
+  }
   const product = new Product({
     // _id:new mongoose.Types.ObjectId('676ba2147d5c97cc4c2be7ba'),
     title: title,
     price: price,
     description: description,
-    imageUrl: imageUrl,
+    imageUrl: imageUrl.path,
     userId: req.user
   });
   product
@@ -33,6 +35,7 @@ exports.postAddProduct = (req, res, next) => {
       res.redirect('/admin/products');
     })
     .catch(err => {
+      console.log('err ----',err)
     const error=new Error(err)
     error.httpStatusCode=500;
     return (next(error))
@@ -69,7 +72,8 @@ exports.postEditProduct = (req, res, next) => {
   const prodId = req.body.productId;
   const updatedTitle = req.body.title;
   const updatedPrice = req.body.price;
-  const updatedImageUrl = req.body.imageUrl;
+  // const imageUrl=req.body.imageUrl
+  const image = req.file;
   const updatedDesc = req.body.description;
 
   Product.findById(prodId)
@@ -81,7 +85,11 @@ exports.postEditProduct = (req, res, next) => {
       product.title = updatedTitle;
       product.price = updatedPrice;
       product.description = updatedDesc;
-      product.imageUrl = updatedImageUrl;
+      if(image){
+        fileHelper.deleteFile(product.imageUrl)
+        product.imageUrl = image.path;
+      }
+    
       return product.save().then(result => {
         console.log('UPDATED PRODUCT!');
         res.redirect('/admin/products');
@@ -94,12 +102,10 @@ exports.postEditProduct = (req, res, next) => {
 exports.getProducts = (req, res, next) => {
   Product.find()
     .then(products => {
-    
       res.render('admin/products', {
         prods: products,
         pageTitle: 'Admin Products',
         path: '/admin/products',
-        
       });
     })
     .catch(err => {
@@ -108,18 +114,28 @@ exports.getProducts = (req, res, next) => {
       return (next(error))
       });
 };
+
 
 exports.postDeleteProduct = (req, res, next) => {
   const prodId = req.body.productId;
 
-  Product.deleteOne({_id:prodId,userId:req.user._id})
-    .then(() => {
-      // console.log('DESTROYED PRODUCT');
-      res.redirect('/admin/products');
-    })
-    .catch(err => {
-      const error=new Error(err)
-      error.httpStatusCode=500;
-      return (next(error))
-      });
+  Product.findById({_id: prodId}).then(product => {
+    if (!product) {
+      return next(new Error("Product Not Found"));
+    }
+    if (product.imageUrl) {
+      fileHelper.deleteFile(product.imageUrl);
+    }
+
+    return Product.deleteOne({_id: prodId, userId: req.user._id});
+  })
+  .then(() => {
+    res.redirect('/admin/products');
+  })
+  .catch(err => {
+    const error = new Error(err);
+    error.httpStatusCode = 500;
+    return next(error);
+  });
 };
+
